Clean up `metabase/lib/urls`
