+++ frontend/src/metabase/lib/urls/admin.ts
@@ -1,0 +1,33 @@
+import { DatabaseId, UserId } from "metabase-types/api";
+
+export function newUser() {
+  return `/admin/people/new`;
+}
+
+export function editUser(userId: UserId) {
+  return `/admin/people/${userId}/edit`;
+}
+
+export function resetPassword(userId: UserId) {
+  return `/admin/people/${userId}/reset`;
+}
+
+export function newUserSuccess(userId: UserId) {
+  return `/admin/people/${userId}/success`;
+}
+
+export function deactivateUser(userId: UserId) {
+  return `/admin/people/${userId}/deactivate`;
+}
+
+export function reactivateUser(userId: UserId) {
+  return `/admin/people/${userId}/reactivate`;
+}
+
+export function newDatabase() {
+  return `/admin/databases/create`;
+}
+
+export function editDatabase(databaseId: DatabaseId) {
+  return `/admin/databases/${databaseId}`;
+}
