+++ frontend/src/metabase/lib/urls/index.ts
@@ -1,0 +1,9 @@
+export * from "./admin";
+export * from "./browse";
+export * from "./collections";
+export * from "./dashboards";
+export * from "./misc";
+export * from "./pulses";
+export * from "./questions";
+export * from "./timelines";
+export * from "./utils";
