+++ frontend/src/metabase/lib/urls/pulses.ts
@@ -1,0 +1,9 @@
+export const newPulse = () => `/pulse/create`;
+
+export function pulse(pulseId: number) {
+  return `/pulse/${pulseId}`;
+}
+
+export function pulseEdit(pulseId: number) {
+  return `/pulse/${pulseId}`;
+}
